refactor: migrate notification API to custom fetch hooks
