+++ src/entities/notification/api/index.ts
@@ -1,5 +1,3 @@
 export * from './endpoint'
 export * from './queryKey'
-export * from './get-notifications'
-export * from './read-notification'
-export * from './post-deviceToken'
+export * from './use-notifications'
